+++ packages/container/angel_container_generator/CHANGELOG.md
@@ -1,2 +1,15 @@
-# 1.0.1
+# Change Log
+
+## 3.0.0-beta.1
+
+* Migrated to support Dart SDK 2.12.x NNBD
+* Updated linter to `package:lints`
+* Updated to use `angel3_` packages
+
+## 2.0.0
+
+* Migrated to work with Dart SDK 2.12.x Non NNBD
+
+## 1.0.1
+
 * Update for `pkg:angel_container@1.0.3`.